Move WalletContext helper funcs to hooks
