+++ src/contexts/WalletContext/hooks/useAddEthereumChain.ts
@@ -3,7 +3,7 @@ import { InjectedProvider } from "types/ethereum";
 import { useLazyChainQuery } from "services/apes";
 import { WalletError } from "errors/errors";
 import { EIPMethods } from "constants/ethereum";
-import toPrefixedHex from "./helpers/toPrefixedHex";
+import toPrefixedHex from "../helpers/toPrefixedHex";
 
 export function useAddEthereumChain() {
   const [getChain] = useLazyChainQuery();
