+++ src/contexts/WalletContext/hooks/useVerifyChain.ts
@@ -1,0 +1,42 @@
+import { useCallback, useEffect } from "react";
+import { Chain } from "types/aws";
+import { useErrorContext } from "contexts/ErrorContext";
+import { WalletDisconnectedError, WrongNetworkError } from "errors/errors";
+import { EXPECTED_NETWORK_TYPE } from "constants/env";
+
+export function useVerifyChain(
+  chain: Chain | undefined,
+  chainError: any,
+  disconnect: () => void
+) {
+  const { handleError } = useErrorContext();
+
+  const handle = useCallback(
+    (error: any) => {
+      handleError(error);
+      try {
+        disconnect();
+      } catch (err) {
+        // when wallet is disconnected, the `disconnect` func is recreated,
+        // causing this hook to rerun and throwing the error below.
+        // We ignore this error and rethrow others
+        if (!(err instanceof WalletDisconnectedError)) {
+          handleError(err);
+        }
+      }
+    },
+    [handleError, disconnect]
+  );
+
+  useEffect(() => {
+    // no active provider === no connected wallet so no need to run hook
+    if (!chain) {
+      return;
+    }
+    if (chainError) {
+      handle(chainError);
+    } else if (chain.network_type !== EXPECTED_NETWORK_TYPE) {
+      handle(new WrongNetworkError());
+    }
+  }, [chain, chainError, handle]);
+}
