+++ src/contexts/WalletContext/useInjectedProvider.ts
@@ -21,7 +21,7 @@ import { WALLET_METADATA } from "./constants";
 import checkXdefiPriority from "./helpers/checkXdefiPriority";
 import { retrieveUserAction, saveUserAction } from "./helpers/prefActions";
 import toPrefixedHex from "./helpers/toPrefixedHex";
-import { useAddEthereumChain } from "./useAddEthereumChain";
+import { useAddEthereumChain } from "./hooks";
 
 const CHAIN_NOT_ADDED_CODE = 4902;
